orchestrator/pyproject.toml
[project]
name = "agent-ui-orchestrator"
version = "0.1.0"
description = "Agent orchestration with WebSocket communication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=14.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "autogen-agentchat>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
